requirements.txt
fastapi
uvicorn
python-dotenv
ibm-watson-machine-learning
pydantic
